Allow AOCL-DA to deal with row major data [CPUPL-5776]
The following changes in this patchset enable this:
1. Move error and options objects into the basic handle class to remove redundant code duplication when dealing with row major data
2. Add option to each handle for row major data (default column major)
3. Add utility functons to convert between row major and update data pointers appropriately
4. Update algorithmic codes to query option and if appropriate convert row major data to column major on entry and back again on exit: future work on algorithm per algorithm basis can improve this
5. Update Python layers to avoid implcit conversion and set option instead and remove need for Fortran style in bindings
6. Update tests and documentation etc
7. Add enums to metrics and basic stats routines as they don't use handles
8. Datastore updates and utility functions

Reviewed-on: http://gerrit-git.amd.com/c/cpulibraries/er/aocl-da/+/1125874
Tested-by: Osi Request <Osi.Request@amd.com>
Tested-by: Llvm Cpu <llvm.cpu@amd.com>

+++ python_interface/python_package/aoclda/examples/knn_ex.py
@@ -31,6 +31,7 @@ import sys
 import numpy as np
 from aoclda.nearest_neighbors import knn_classifier
 
+
 def knn_classifier_example():
     """
     k-Nearest Neighbors classification
@@ -42,13 +43,13 @@ def knn_classifier_example():
                         [-3, -2, -1],
                         [1, 3, 1],
                         [2, 5, 1],
-                        [3, -1, 2]])
+                        [3, -1, 2]], dtype=np.float64)
 
     y_train = np.array([1, 2, 0, 1, 2, 2])
 
-    x_test = np.array([[-2 , 2, 3],
+    x_test = np.array([[-2, 2, 3],
                        [-1, -2, -1],
-                       [2, 1, -3]])
+                       [2, 1, -3]], dtype=np.float64)
 
     print("\nk-nearest neighbors for a small data matrix\n")
     knn = knn_classifier()
@@ -72,8 +73,8 @@ def knn_classifier_example():
     expected_ind = np.array([[1, 0, 3],
                              [2, 0, 1],
                              [3, 5, 4]])
-    expected_dist = np.array([[3.        , 3.31662479, 3.74165739 ],
-                              [2.        , 3.16227766, 4.24264069 ],
+    expected_dist = np.array([[3., 3.31662479, 3.74165739],
+                              [2., 3.16227766, 4.24264069],
                               [4.58257569, 5.47722558, 5.65685425]])
     expected_proba = np.array([[0.2, 0.4, 0.4],
                                [0.2, 0.4, 0.4],
@@ -96,6 +97,7 @@ def knn_classifier_example():
         sys.exit(1)
 
     print("\nk-nearest neighbors successfully computed\n")
+
 
 if __name__ == "__main__":
     try:
