+++ python_interface/tests/sklearnAPI/knn_sklearn_test.py
@@ -62,8 +62,8 @@ def test_knn_classifier(precision, weights, metric, n_neigh_constructor, n_neigh
     # patch and import scikit-learn
     skpatch()
     from sklearn import neighbors
-    knn_da = neighbors.KNeighborsClassifier(weights=weights, 
-                                            n_neighbors=n_neigh_constructor, 
+    knn_da = neighbors.KNeighborsClassifier(weights=weights,
+                                            n_neighbors=n_neigh_constructor,
                                             metric=metric)
     knn_da.fit(x_train, y_train)
     da_dist, da_ind = knn_da.kneighbors(x_test, n_neighbors=n_neigh_kneighbors, return_distance=True)
@@ -122,10 +122,10 @@ def test_knn_errors():
     with pytest.raises(ValueError):
         knn = neighbors.KNeighborsClassifier(metric = "manhattan")
 
-    x_train = np.array([[1, 1, 1], [2, 2, 2], [3, 3, 3]])
-    y_train = np.array([[1, 2, 3]])
-    x_test = np.array([[1, 2, 3], [3, 2, 1]])
-    y_test = np.array([[1, 1]])
+    x_train = np.array([[1, 1, 1], [2, 2, 2], [3, 3, 3]], dtype=np.float64)
+    y_train = np.array([[1, 2, 3]], dtype=np.float64)
+    x_test = np.array([[1, 2, 3], [3, 2, 1]], dtype=np.float64)
+    y_test = np.array([[1, 1]], dtype=np.float64)
     knn = neighbors.KNeighborsClassifier()
     knn.fit(x_train, y_train)
     with pytest.raises(RuntimeError):
